+++ inst/doc/derivmkts-vignette.R
@@ -37,6 +37,12 @@ bsput(s, c(80, 100, 120), v, r, tt, d)
 uoput(s, c(80, 100, 120), v, r, tt, d, H)
 
 ## ------------------------------------------------------------------------
+s <- 100; k <- 100; r <- 0.08; v <- 0.30; tt <- 2; d <- 0.04
+callperpetual(s, c(95, 100, 105), v, r, d)
+callperpetual(s, c(95, 100, 105), v, r, d, showbarrier=TRUE)
+
+
+## ------------------------------------------------------------------------
 H <- 105
 greeks(uicall(s, k, v, r, tt, d, H))
 
@@ -45,6 +51,8 @@ greeks(uicall(s, k, v, r, tt, d, H))
 powercontract <- function(s, v, r, tt, d, a) {
     price <- exp(-r*tt)*s^a* exp((a*(r-d) + 1/2*a*(a-1)*v^2)*tt)
 }
+
+## ------------------------------------------------------------------------
 greeks(powercontract(s=40, v=.08, r=0.08, tt=0.25, d=0, a=2))
 
 ## ------------------------------------------------------------------------
@@ -60,18 +68,17 @@ x <- greeks(bullspread(sseq, .3, .08, 1, 0, k1=40, k2=45))
 plot(sseq, x['Gamma',], type='l')
 
 
-## ----allgreeks, fig.cap='All option Greeks, plotted using bsopt', fig.width=7.5, fig.height=6.5----
+## ----allgreeks, fig.cap='All option Greeks, computed using the greeks() function', fig.width=7.5, fig.height=6.5----
 k <- 100; r <- 0.08; v <- 0.30; tt <- 2; d <- 0
 S <- seq(.5, 250, by=.5)
-cgreeks <- greeks(bscall(S, k, v, r, tt, d))
-pgreeks <- greeks(bsput(S, k, v, r, tt, d))
-optlbl <-  c('Call', 'Put')
-y <- list(cgreeks, pgreeks)
+Call <- greeks(bscall(S, k, v, r, tt, d))
+Put <- greeks(bsput(S, k, v, r, tt, d))
+y <- list(Call=Call, Put=Put)
 par(mfrow=c(4, 4))  ## create a 4x4 plot
 par(mar=c(2,2,2,2))
-for (i in 1:length(y)) {
+for (i in names(y)) {
     for (j in rownames(y[[i]])) {  ## loop over greeks
-        plot(S, y[[i]][j, ], main=paste(optlbl[i], j), ylab=j, type='l')
+        plot(S, y[[i]][j, ], main=paste(i, j), ylab=j, type='l')
     }
 }
 
@@ -84,8 +91,10 @@ binomopt(s, k, v, r, tt, d, nstep=3, putopt=TRUE)
 binomopt(s, k, v, r, tt, d, nstep=3, returntrees=TRUE, putopt=TRUE)
 
 ## ------------------------------------------------------------------------
-geomavgprice(s, k, v, r, tt, d, 3)
-geomavgstrike(s, k, v, r, tt, d, 3)
+s <- 100; k <- 100; r <- 0.08; v <- 0.30; tt <- 2; d <- 0.03; m <- 3
+geomavgpricecall(s, 98:102, v, r, tt, d, m)
+geomavgpricecall(s, 98:102, v, r, tt, d, m, cont=TRUE)
+geomavgstrikecall(s, k, v, r, tt, d, m)
 
 
 ## ------------------------------------------------------------------------
@@ -129,4 +138,128 @@ binomplot(s, k, v, r, tt, d, nstep=40, american=TRUE)
 ## ----binomplot4, fig.cap="Binomial plot when nstep is 40 using the argument ylimval to focus on a subset.\\label{fig:binomplot4}"----
 d <- 0.06
 binomplot(s, k, v, r, tt, d, nstep=40, american=TRUE, ylimval=c(75, 225))
+
+## ------------------------------------------------------------------------
+f = function(a, b, k) a*b + k
+f(3, 5, 1) 
+f(1:5, 5, 1) 
+f(1:6, 1:2, 1) 
+
+## ------------------------------------------------------------------------
+cond1 <- function(a, b, k) {
+    if (a > b) {
+        a*b + k
+    } else {
+        k
+    }
+}
+cond1(5, 3, 1)
+cond1(5, 7, 1)
+cond1(3:7, 5, 1)
+
+
+## ------------------------------------------------------------------------
+cond2 <- function(a, b, k) {
+    ifelse(a > b, 
+           a*b + k,  
+           k
+           )
+}
+cond2(5, 3, 1)
+cond2(5, 7, 1)
+cond2(3:7, 5, 1)
+
+
+## ------------------------------------------------------------------------
+cond2(5, 7, 1:3)
+
+## ------------------------------------------------------------------------
+cond2b <- function(a, b, k) {
+    agtb <- (a > b)
+    agtb*(a*b + k) + (1-agtb)*k
+}
+
+cond2b(5, 3, 1)
+cond2b(5, 7, 1)
+cond2b(3:7, 5, 1)
+cond2b(5, 7, 1:3)
+
+
+## ------------------------------------------------------------------------
+cond2c <- function(a, b, k) {
+    tmp <- data.frame(a, b, k)
+    for (i in names(tmp)) assign(i, tmp[[i]])
+    ifelse(a > b, 
+           a*b + k,  
+           k
+           )
+}
+
+cond2c(5, 3, 1)
+cond2c(5, 7, 1)
+cond2c(3:7, 5, 1)
+cond2c(5, 7, 1:3)
+
+
+## ------------------------------------------------------------------------
+vectorizeinputs <- function(e) {
+    ## e is the result of match.call() in the calling function
+    e[[1]] <- NULL
+    e <- as.data.frame(as.list(e))
+    for (i in names(e)) assign(i, eval(e[[i]]),
+                               envir=parent.frame())
+}
+
+## ------------------------------------------------------------------------
+cond3 <- function(a, b, k) { 
+    vectorizeinputs(match.call())
+    ifelse(a > b, a*b + k, k)
+}
+cond3(5, 7, 1:3)
+cond3(3:7, 5, 1)
+cond3(3:7, 5, 1:5)
+cond3(k=1:5, 3:7, 5)
+
+
+## ------------------------------------------------------------------------
+cond4 <- function(a, b, k, multby2=TRUE) {
+    vectorizeinputs(match.call())
+    ifelse(multby2, 
+           2*(a*b + k), 
+           a*b + k
+           )
+}
+cond4(5, 7, 1:3)
+cond4(3:7, 5, 1)
+cond4(3:7, 5, 1:5)
+cond4(k=1:5, 3:7, 5)
+
+
+## ------------------------------------------------------------------------
+vectorizeinputs2 <- function(e) {
+    funcname <- e[[1]]
+    fvals <- formals(eval(funcname))
+    fnames <- names(fvals)
+    e[[1]] <- NULL
+    e <- as.data.frame(as.list(e))
+    implicit <- setdiff(fnames, names(e))
+    if (length(implicit) > 0) e <- data.frame(e, fvals[implicit])
+    for (i in names(e)) assign(i, eval(e[[i]]),
+                               envir=parent.frame())
+}
+
+cond5 <- function(a, b, k, multby2=TRUE, altmult=1) {
+    vectorizeinputs2(match.call())
+    ifelse(multby2, 
+           2*(a*b + k), 
+           altmult*(a*b + k)
+           )
+}
+cond5(5, 7, 1:3)
+cond5(3:7, 5, 1)
+cond5(3:7, 5, 1:5)
+cond5(k=1:5, 3:7, 5)
+cond5(k=1:5, 3:7, 5, multby2=FALSE)
+cond5(k=1:5, 3:7, 5, multby2=FALSE, altmult=5)
+